heatmap.py save all maps

+++ src/heatMap.py
@@ -1,3 +1,4 @@
+import os
 import pandas
 import numpy as np
 import csv
@@ -9,25 +10,21 @@ import matplotlib.pyplot as plt
 
 
 goalsDF = csvToDF('../data/goals.csv')
-print(goalsDf)
+
 goalsDF = goalsDF.drop(goalsDF.loc[goalsDF['x'] == 'NA'].index)
 goalsDF = goalsDF.drop(goalsDF.loc[goalsDF['x'] == 'x'].index)
 goalsDF = goalsDF.drop(goalsDF.loc[goalsDF['secondaryType'] == 'x'].index)
+goalsDF = goalsDF.drop(goalsDF.loc[goalsDF['secondaryType'] == 'NA'].index)
 
 goalsDF.loc[:,'x'] = goalsDF.x.astype(np.float)
 goalsDF.loc[:,'y'] = goalsDF.y.astype(np.float)
 
 for goalType in (goalsDF['secondaryType'].unique()):
-    goalsDF = goalsDF.loc[goalsDF['secondaryType'] == goalType]
-
-    heatmapGoalsDF = pandas.DataFrame(np.array([]))
-    heatmapGoalsDF['x'] = goalsDF['x']
-    heatmapGoalsDF['y'] = goalsDF['y']
-    heatmapGoalsDF = heatmapGoalsDF.drop(0, axis=1)
-
-    sns.jointplot(x='x', y='y', data=heatmapGoalsDF, kind='kde', color='r', xlim=(-100,100), ylim=(-42.5,42.5))
+    sns.jointplot(x='x', y='y', data=goalsDF.loc[goalsDF['secondaryType'] == goalType],
+                  kind='kde', color='r', xlim=(-100,100), ylim=(-42.5,42.5))
 
     checkAndMakeImgFolder(heatmap=True)
 
-    os.remove('../img/heatmaps/jointplot-' + '-'.join(goalType.lower().split(' ')) + '-heatmap.png')
+    if os.path.exists('../img/heatmaps/jointplot-' + '-'.join(goalType.lower().split(' ')) + '-heatmap.png'):
+        os.remove('../img/heatmaps/jointplot-' + '-'.join(goalType.lower().split(' ')) + '-heatmap.png')
     plt.savefig('../img/heatmaps/jointplot-' + '-'.join(goalType.lower().split(' ')) + '-heatmap.png', bbox_inches='tight')
